+++ lldb/source/Expression/ClangFunction.cpp
@@ -173,7 +173,7 @@ ClangFunction::CompileFunction (Stream &errors)
             }
             else
             {   
-                errors.Printf("Could not determine type of input value %" PRId64 ".", i);
+                errors.Printf("Could not determine type of input value %" PRIu64 ".", (uint64_t)i);
                 return 1;
             }
         }
@@ -344,7 +344,7 @@ ClangFunction::WriteFunctionArguments (ExecutionContext &exe_ctx,
     size_t num_args = arg_values.GetSize();
     if (num_args != m_arg_values.GetSize())
     {
-        errors.Printf ("Wrong number of arguments - was: %" PRId64 " should be: %" PRId64 "", num_args, m_arg_values.GetSize());
+        errors.Printf ("Wrong number of arguments - was: %" PRIu64 " should be: %" PRIu64 "", (uint64_t)num_args, (uint64_t)m_arg_values.GetSize());
         return false;
     }
